added subscription preprocessor
