+++ packages/houdini-common/src/config.ts
@@ -20,7 +20,6 @@ export class Config {
 	sourceGlob: string
 	quiet: boolean
 	verifyHash: boolean
-	mode: string
 
 	constructor({
 		schema,
@@ -29,7 +28,6 @@ export class Config {
 		quiet = false,
 		verifyHash,
 		filepath,
-		mode = 'sapper',
 	}: ConfigFile & { filepath: string }) {
 		// make sure we got some kind of schema
 		if (!schema && !schemaPath) {
@@ -50,16 +48,12 @@ export class Config {
 		this.sourceGlob = sourceGlob
 		this.quiet = quiet
 		this.verifyHash = typeof verifyHash === 'undefined' ? true : verifyHash
-		this.mode = mode
 
 		// if we are building a sapper project, we want to put the runtime in
 		// src/node_modules so that we can access @sapper/app and interact
 		// with the application stores directly
 		const rootDir = path.dirname(filepath)
-		this.rootDir =
-			mode === 'sapper'
-				? path.join(rootDir, 'src', 'node_modules', '$houdini')
-				: path.join(rootDir, '.houdini')
+		this.rootDir = path.join(rootDir, 'src', 'node_modules', '$houdini')
 	}
 
 	/*
